Migrate to frontmatter-markdown-loader 3.X
The loader is no longer returning strings with the bodies of the render
functions. Instead, it is returning the proper functions. These are not
serializable and cannot be returned from asyncData according to:
https://github.com/nuxt/nuxt.js/issues/4424#issuecomment-442179494

To recover the functionallity, the commit transforms the functions into
serializable versions (a stringified version) of them.